refactor: move uri to domain package
