+++ src/jsTest/kotlin/com/abogomazov/application.domain/UriTest.kt
@@ -1,4 +1,4 @@
-package com.abogomazov
+package com.abogomazov.application.domain
 
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
